Move actions.js into actions directory
