+++ client/components/MessageBoxes.jsx
@@ -1,7 +1,7 @@
 import React, { Component, PropTypes } from 'react';
 import MessageBox from './MessageBox.jsx'
 import { connect } from 'react-redux';
-import { checkAuthentication } from '../actions.js';
+import { checkAuthentication } from './../actions';
 
 // console.log('checkAuthentication', checkAuthentication);
