+++ client/components/PopularItems.jsx
@@ -2,7 +2,7 @@ import React, { Component, PropTypes } from 'react';
 import ReactDOM from 'react-dom';
 import { connect } from 'react-redux';
 import $ from 'jquery';
-import { checkAuthentication } from '../actions.js';
+import { checkAuthentication } from './../actions';
 import ItemEntry from './ItemEntry.jsx';
 
 class PopularItems extends Component {
