+++ client/components/Profile.jsx
@@ -1,7 +1,7 @@
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
 
-import { getProfile } from '../actions.js';
+import { getProfile } from './../actions';
 import Feedback from './Feedback.jsx';
 import ProfileEntry from './ProfileEntry.jsx';
